split up runtime inspector tests
